Misc
- Make the amount of essence in a soulbook that is required to insert it into a summoner configurable
- Change the essence of murky orbs and soulbooks into integers
- Make drops configurable by whether the entity they're dropped from was summoned or spawned

+++ src/main/java/yuudaari/soulus/common/config/Config.java
@@ -34,7 +34,7 @@ public class Config {
 	public int boneChunkParticleCount = 3;
 	public List<EssenceConfig> essences = EssenceConfig.getDefaultCreatureConfigs();
 
-	public byte getSoulbookQuantity(String essenceType) {
+	public int getSoulbookQuantity(String essenceType) {
 		for (EssenceConfig config : essences) {
 			if (config.essence.equals(essenceType)) {
 				return config.soulbookQuantity;
